Update from April 3, 2006
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@647 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/common/state/PickPointInfo.h
@@ -29,8 +29,8 @@ typedef struct PickPointInfo
 {
     bool    validPick;
     bool    intersectionOnly;
-    float   rayPt1[3];
-    float   rayPt2[3];
+    double  rayPt1[3];
+    double  rayPt2[3];
     void   *callbackData;
 } PICK_POINT_INFO;
