+++ src/components/Database/Database/avtSourceFromDatabase.h
@@ -112,9 +112,9 @@ class DATABASE_API avtSourceFromDatabase : public avtTerminatingDatasetSource
     void                DatabaseProgress(int, int, const char *);
     virtual void        Query(PickAttributes *);
     virtual bool        QueryCoords(const std::string &, const int, const int,
-                                    const int, float[3], const bool, const bool, const char *mn = NULL);
+                                    const int, double[3], const bool, const bool, const char *mn = NULL);
     virtual bool        FindElementForPoint(const char *, const int, 
-                            const int, const char *, float[3], int &);
+                            const int, const char *, double[3], int &);
     virtual void        GetDomainName(const std::string &, const int, 
                             const int, std::string &);
